+++ modules/router/src/index.ts
@@ -107,7 +107,7 @@ const evts: EventCallbackConfig = {
 
 const signer = new ChannelSigner(Wallet.fromMnemonic(config.mnemonic).privateKey);
 
-const logger = pino({ name: signer.publicIdentifier });
+const logger = pino({ name: signer.publicIdentifier, level: config.logLevel });
 logger.info("Loaded config from environment");
 const server = fastify({
   logger,