+++ modules/router/src/listener.ts
@@ -49,6 +49,7 @@ import {
 } from "./metrics";
 import { calculateFeeAmount } from "./services/fees";
 import { QuoteError } from "./errors";
+import { inProgressCreations } from "./services/creationQueue";
 
 const config = getConfig();
 
@@ -57,15 +58,6 @@ const ajv = new Ajv();
 export type ChainJsonProviders = {
   [k: string]: JsonRpcProvider;
 };
-
-// Used to track all the transfers we are forwarding in memory
-// so that when router is handling transfers they may have dropped,
-// they do not double spend. I.e. sender creates transfer and goes
-// offline. Router starts forwarding to receiver, and while this is
-// happening sender comes back online. Without tracking the in-progress
-// forwards, the transfer would be double created with the receiver via
-// the handleIsAlive fn
-export const inProgressCreations: { [channelAddr: string]: string[] } = {};
 
 export async function setupListeners(
   routerSigner: IChannelSigner,