+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- \[router\] Create queues for transfer creation races
 - \[engine, types\] Make timeout parameter optional for setup, conditional transfer, and withdraw
 
 ## 0.2.1-beta.2